requirements.txt
biopython==1.71
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
Werkzeug==0.14.1
